+++ Realm/Realm.Sync/Properties/AssemblyInfo.cs
@@ -22,3 +22,5 @@ using System.Runtime.CompilerServices;
 [assembly: AssemblyTitle("Realm.Sync")]
 [assembly: InternalsVisibleTo("Tests.XamarinAndroid")]
 [assembly: InternalsVisibleTo("Tests.XamarinIOS")]
+[assembly: InternalsVisibleTo("Tests.iOS")]
+[assembly: InternalsVisibleTo("Tests.Android")]