+++ Realm/Realm/Properties/AssemblyInfo.cs
@@ -26,3 +26,5 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Tests.Win32")]
 [assembly: InternalsVisibleTo("Tests.XamarinAndroid")]
 [assembly: InternalsVisibleTo("Tests.XamarinIOS")]
+[assembly: InternalsVisibleTo("Tests.iOS")]
+[assembly: InternalsVisibleTo("Tests.Android")]